% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_and_plot_approximation_ratio.R
\name{calculate_and_plot_approximation_ratio}
\alias{calculate_and_plot_approximation_ratio}
\title{Calculate and Plot Approximation Ratio}
\usage{
calculate_and_plot_approximation_ratio(
  solutions,
  exact_solutions,
  method_names,
  save_format = NULL,
  file_name = "approximation_ratio_plot"
)
}
\arguments{
\item{solutions}{A list of numeric vectors representing the solutions found. Each list corresponds to an instance.}

\item{exact_solutions}{A numeric vector representing the exact solutions. The length of the vector should match the length of the `solutions` list.}

\item{method_names}{A vector of strings representing the names of the different methods.}

\item{save_format}{The format in which the plot should be saved: "pdf", "png" or NULL. By default, the plot is not saved.}

\item{file_name}{The name of the file if the plot is to be saved. Default is "approximation_ratio_plot".}
}
\value{
A list with two elements: `average_ratios` and `plot`.
}
\description{
This function calculates the approximation ratios for multiple instances and methods, and creates a boxplot.
}
\examples{
solutions <- list(
  method1 = list(c(100, 110, 120), c(200, 210, 220)),
  method2 = list(c(105, 115, 125), c(205, 215, 225))
)
exact_solutions <- c(100, 200)
method_names <- c("Method 1", "Method 2")
calculate_and_plot_approximation_ratio(solutions, exact_solutions, method_names)
}
