% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R, R/운동.R
\name{exe}
\alias{exe}
\alias{exe}
\title{generate exercise data}
\usage{
exe(ex, input, rawdata = NULL, summary = F)

exe(ex, input, rawdata = NULL, summary = F)
}
\value{
exercise data

exercise data
}
\description{
generate exercise data

generate exercise data
}
\examples{
data<-exe('kettlebell_swing',input=c(30,30,30,30))
exdata<-exe('kettlebell_swing',c(30,30,30),rawdata=exdata)
data<-exe('kettlebell_swing',input=c(30,30,30,30))
exdata<-exe('kettlebell_swing',c(30,30,30),rawdata=exdata)
}
