% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\docType{data}
\name{exercise}
\alias{exercise}
\title{Exercise Description}
\format{
A data frame with 973 rows and 15 variables:
\describe{
\item{Age}{Age of the gym member.}
\item{Gender}{Gender of the gym member (Male or Female).}
\item{Weight (kg)}{Member’s weight in kilograms.}
\item{Height (m)}{Member’s height in meters.}
\item{Max_BPM}{Maximum heart rate (beats per minute) during workout sessions.}
\item{Avg_BPM}{Average heart rate during workout sessions.}
\item{Resting_BPM}{Heart rate at rest before workout.}
\item{Session_Duration (hours)}{Duration of each workout session in hours.}
\item{Calories_Burned}{Total calories burned during each session.}
\item{Workout_Type}{Type of workout performed (e.g., Cardio, Strength, Yoga, HIIT).}
\item{Fat_Percentage}{Body fat percentage of the member.}
\item{Water_Intake (liters)}{Daily water intake during workouts.}
\item{Workout_Frequency (days/week)}{Number of workout sessions per week.}
\item{Experience_Level}{Level of experience, from beginner (1) to expert (3).}
\item{BMI}{Body Mass Index, calculated from height and weight.}
}
}
\source{
Kaggle (see link above)
}
\usage{
exercise
}
\description{
Info available here: https://www.kaggle.com/datasets/valakhorasani/gym-members-exercise-dataset.
}
\keyword{datasets}
