% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_function.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing values in a dataset}
\usage{
impute_missing(data, method)
}
\arguments{
\item{data}{A data frame containing missing values.}

\item{method}{The imputation method to be used (e.g., "mean", "median", "knn").}
}
\value{
A data frame with missing values imputed.
}
\description{
This function replaces missing values in a dataset with appropriate imputed values.
}
\examples{
data <- data.frame(x = c(1, 2, NA, 4, 5), y = c(NA, 2, 3, 4, NA))
imputed_data <- impute_missing(data, method = "mean")
imputed_data
}
