% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data}
\alias{plot_data}
\title{Plot a dataset}
\usage{
plot_data(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The name of the x variable.}

\item{y}{The name of the y variable.}
}
\description{
This function creates a basic plot of the given dataset.
}
\examples{
plot_data(iris, "Sepal.Length", "Sepal.Width")
}
