% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variables.R
\name{plot_variables}
\alias{plot_variables}
\title{Plot Variables}
\arguments{
\item{data}{A data frame containing the data.}

\item{var1}{A string representing the first variable.}

\item{var2}{A string representing the second variable.}

\item{save_format}{The format in which the plot should be saved: "pdf", "png" or NULL. By default, the plot is not saved.}

\item{file_name}{The name of the file if the plot is to be saved. Default is "scatter_plot".}
}
\value{
A ggplot2 object.
}
\description{
This function creates a scatter plot between two variables in a given dataset.
}
\examples{
data <- data.frame(x = rnorm(100), y = rnorm(100))
plot_variables(data, "x", "y")
}
