% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semabroadessential.R
\name{semabroadessential}
\alias{semabroadessential}
\title{Essential places of your semester abroad}
\usage{
semabroadessential(city)
}
\arguments{
\item{city}{The place you are going to study in. Names may have to be written completely in small letters.}
}
\value{
A map of that place with essential amenities.
}
\description{
Essential places of your semester abroad
}
\examples{
semabroadessential("cape town")

}
