% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semabroadhelp.R
\name{semabroadhelp}
\alias{semabroadhelp}
\title{Possible places of interest}
\usage{
semabroadhelp()
}
\value{
A couple of categories (keys) and features (values) to be used in conjunction with semabroadmap().
}
\description{
Possible places of interest
}
\examples{
semabroadhelp()
}
