% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semabroadmap.R
\name{semabroadmap}
\alias{semabroadmap}
\title{The city you will study in, for you to explore}
\usage{
semabroadmap(city, category, type)
}
\arguments{
\item{city}{The city you're going to study in.}

\item{category}{A category of places. See semabroadhelp().}

\item{type}{A type of place included in the category. See semabroadhelp().}
}
\value{
A map showing places of interest as indicated within the given city.
}
\description{
The city you will study in, for you to explore
}
\examples{
semabroadmap("cape town", "tourism", "attraction")
}
