% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Summarize a dataset}
\usage{
summary_statistics(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A summary of the dataset.
}
\description{
This function generates summary statistics for a given dataset.
}
\examples{
summary_statistics(iris)
}
